project/LosSilksongs/userservice/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.7
alembic==1.11.1
pydantic==1.10.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
python-dotenv==1.1.0
email-validator==2.0.0.post2
PyYAML==6.0.2
python-multipart==0.0.6